Debit and Credit fields are altered to Float from Integer.

+++ mysite/tracker/models.py
@@ -4,9 +4,9 @@ from django.db import models
 
 class Account(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    debit = models.IntegerField(default=0)
-    credit = models.IntegerField(default=0)
-    balance = models.IntegerField(default=0)
+    debit = models.FloatField(default=0)
+    credit = models.FloatField(default=0)
+    balance = models.FloatField(default=0)
 
     def __str__(self):
         return str(self.user)
@@ -14,8 +14,8 @@ class Account(models.Model):
 
 class Transaction(models.Model):
     user = models.ForeignKey(Account, on_delete=models.CASCADE)
-    debit = models.IntegerField(default=0)
-    credit = models.IntegerField(default=0)
+    debit = models.FloatField(default=0)
+    credit = models.FloatField(default=0)
 
     def __str__(self):
         return str(self.debit - self.credit)
